refactored SearchButton and RandomButton to include a new scrape function imported from utils
